+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod client;
 pub mod codec;
 pub mod server;
 pub mod status;